% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrackData.R
\name{duttSim}
\alias{duttSim}
\title{Simulate user tracked menstrual cycle data for an individual using the dutt method.}
\usage{
duttSim(i, skipProb, maxCycles, trueBetas, trueGammas, overlap, xCovF, zCovF)
}
\arguments{
\item{i}{Individual identifier. Character, numeric or integer.}

\item{skipProb}{Vector of probabilities for number of true cycles per tracked cycle. For
example, (.7, .2, .1) means that 70\% of observed cycles will contain one true cycle, 20\%
will contain 2 true cycles and 10\% will contain 3 true cycles.}

\item{maxCycles}{Maximum number of true cycles per tracked cycle. Ignored.}
}
\value{
A data.frame with columns 'Individual', 'TrackedCycles', 'NumTrue',
'LogMean', 'LogPrec'.
}
\description{
This function generates synthetic data for user tracked menstrual cycles for a
single individual using the dutt method.
}
\seealso{
\code{\link{simTrackData}}
}
