% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrackMCMC.R
\name{gibbsStep}
\alias{gibbsStep}
\title{Perform one sampling step for our skipTrackMCMC model}
\usage{
gibbsStep(ijDat, iDat, mu, rho, pi, priorAlphas, indFirst)
}
\arguments{
\item{ijDat}{A data.frame containing all data at the level of individual-observation (ij),
with columns: Individual, ys, cs.}

\item{iDat}{A data.frame containing all data at the level of individual (i),
with columns: Individual, mus, taus.}

\item{mu}{Current value of the global parameter mu.}

\item{rho}{Current value of the global parameter rho.}

\item{pi}{Current value of the global parameter pi.}

\item{priorAlphas}{Vector of prior alpha values for updating pi.}

\item{indFirst}{Logical vector indicating the first instance of each individual in the full data.}
}
\value{
A list containing the updated data and parameters:
\describe{
\item{iDat}{A data.frame with updated parameters at the individual level: Individual, mus, taus.}
\item{ijDat}{A data.frame with updated parameters at the individual-observation level: Individual, ys, cs.}
\item{mu}{Updated value of the global parameter mu.}
\item{rho}{Updated value of the global parameter rho.}
\item{pi}{Updated value of the global parameter pi.}
\item{priorAlphas}{Unchanged vector of prior alpha values for updating pi.}
\item{indFirst}{Unchanged logical vector indFirst}
}
}
\description{
This function updates parameters at three levels: global (\code{mu}, \code{rho}, \code{pi}),
individual (\code{mus}, \code{taus}), and individual-observation (\code{ys}, \code{cs}).
}
\seealso{
\code{\link{postMu}}, \code{\link{postRho}}, \code{\link{postPi}},
\code{\link{postMui}}, \code{\link{postTaui}}, \code{\link{postCij}}
}
\keyword{gibbs}
\keyword{hierarchical}
\keyword{model}
\keyword{parameters}
\keyword{sampling}
\keyword{update}
