% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liInference.R
\name{liki}
\alias{liki}
\title{Calculate the likelihood of parameters for an individual's cycle lengths}
\usage{
liki(di, lambdai, pii, S)
}
\arguments{
\item{di}{Numeric vector representing individual data.}

\item{lambdai}{Numeric value representing the individuals cycle length mean parameter}

\item{pii}{Numeric value representing the individuals probability of skipping tracking a cycle.}

\item{S}{Maximum number of allowed skips in the model.}
}
\value{
Numeric value representing the likelihood of the parameters for the data.
}
\description{
This function calculates the likelihood of the parameters for a given individual's data.
It is adapted from a formula provided in Li et al. (2022)
}
\details{
The function calculates the likelihood for an individuals data based on
the formula provided in the Supplementary Materials by Li et al. (2022).
}
\references{
Li, Kathy, et al. "A predictive model for next cycle start date that accounts for
adherence in menstrual self-tracking." Journal of the American Medical
Informatics Association 29.1 (2022): 3-11.
}
