% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{liSim}
\alias{liSim}
\title{Simulate user tracked menstrual cycle data for an individual using the li model.}
\usage{
liSim(i, skipProb, maxCycles, trueBetas, trueGammas = NULL)
}
\arguments{
\item{i}{Individual identifier. Character, numeric or integer.}

\item{skipProb}{Vector, ignored for this model.}

\item{maxCycles}{Integer, Maximum possible number of true cycles per tracked cycle.}

\item{trueBetas}{Optional. True values for generated mean regression coefficients.}

\item{trueGammas}{NULL, left for consistency. Will throw error if specified.}
}
\value{
A data.frame with columns 'Individual', 'TrackedCycles',
'NumTrue', 'Mean', 'SkipProb', 'Z0', 'Beta0', 'Gamma0', 'X0', ..., 'Xn' where n = length(trueBetas).
}
\description{
This function generates synthetic data for user tracked menstrual cycles for a
single individual using the li model. For this model Beta0 = log(30), and Gamma0 doesn't really make sense.
}
\seealso{
\code{\link{skipTrack.simulate}}
}
