% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.skipTrack.model}
\alias{plot.skipTrack.model}
\title{Plot skipTrack.model objects}
\usage{
\method{plot}{skipTrack.model}(x, ...)
}
\arguments{
\item{x}{skipTrack.model object from the function skipTrack.fit}

\item{...}{Needed for S3 consistency}
}
\value{
Invisible NULL. Prints plots from skipTrack.visualize
}
\description{
Plot skipTrack.model objects
}
\examples{
#Simulated data
simDat <- skipTrack.simulate(n = 100, skipProb = c(.7, .2, .1))

#Run model fit (should typically run with much more than 50 reps)
modFit <- skipTrack.fit(Y = simDat$Y, cluster = simDat$cluster, chains = 2, reps = 50)
plot(modFit)

}
