% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postB}
\alias{postB}
\title{Sample a value from the full conditional posterior of b_i}
\usage{
postB(taui, mi, XiBeta, rho)
}
\arguments{
\item{taui}{Numeric, tau_i for individual i}

\item{mi}{Numeric vector, length equal to the number of cycles contributed by individual i}

\item{XiBeta}{Numeric vector, equal to t(X_i) \%*\% Beta}

\item{rho}{Numeric, rho value for prior precision of b_i}
}
\value{
Numeric value, single draw of bi
}
\description{
Sample a value from the full conditional posterior of b_i
}
