% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postMu}
\alias{postMu}
\title{Sample a value from the full conditional posterior of mu}
\usage{
postMu(muI, rho, priorMean = log(30), priorPre = 0.1)
}
\arguments{
\item{muI}{Numeric vector, log of individuals mean values.}

\item{rho}{Numeric > 0, a sampled precision of the mu_i values}

\item{priorMean}{Numeric, prior mean of mu, default is log(30) to match common menstrual cycle lengths}

\item{priorPre}{Numeric > 0, prior precision of mu, default is 1 to have little influence}
}
\value{
Numeric
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_{ij}), tau_i). The prior for mu_i
is given as N(mu, rho). This function draws from the conditional posterior of mu, given
that the prior on mu is N(priorMean, priorPre).
}
\examples{
mui <- rep(log(31), 10)
postMu(mui, 10)
}
