% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postMuiOLD}
\alias{postMuiOLD}
\title{Sample a value from the full conditional posterior of mu_i (OLD VERSION)}
\usage{
postMuiOLD(yij, cij, taui, mu, rho)
}
\arguments{
\item{yij}{Numeric vector, cycle lengths for a single individual}

\item{cij}{Positive Integer vector, a sampled vector of length(yij) where the corresponding
values in cij indicate a sampled number of TRUE cycles in each cycle length given by yij}

\item{taui}{Numeric > 0, A sampled precision for the yijs}

\item{mu}{Numeric, sampled prior mean of mu_i}

\item{rho}{Numeric > 0, sampled prior precision of mu_i}
}
\value{
Numeric
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_{ij}), tau_i). The prior for mu_i
is given as N(mu, rho). This function draws from the conditional posterior of mu_i.
}
