% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postPhi}
\alias{postPhi}
\title{Sample a value from the full conditional posterior of phi}
\usage{
postPhi(taui, thetai, currentPhi, rhoPhi = 1000)
}
\arguments{
\item{taui}{Numeric vector, individuals precisions.}

\item{thetai}{Numeric vector. individuals precisions means (estimate)}

\item{oldPhi}{Previous draw of phi}

\item{nu0}{Shape parameter of prior on phi}

\item{delta0}{Rate parameter of prior on phi}
}
\value{
Numeric
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_{ij}), tau_i). The prior for tau_i
is given as Gamma(thetai*phi, phi). This function draws from the conditional posterior of phi,
given that the prior on phi is gamma(nu0, delta0). The function calculates individual shape
parameters given the previous phi.
Note that we parameterize with RATE, not SCALE.
}
