% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postPi}
\alias{postPi}
\title{Sample a value from the full conditional posterior of pi}
\usage{
postPi(ci, priorAlphas)
}
\arguments{
\item{ci}{Integer vector, all of the sampled cij values for all individuals}

\item{priorAlphas}{Numeric vector, prior dirichlet parameters for pi}
}
\value{
Numeric vector
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_ij), tau_i) The prior for
c_ij is a categorical prior with category probabilities pi1, ..., pik, and c_ij can
take values 1, ..., k where k is the length of pi. This function samples from the
posterior of pi = pi1, ..., pik, assuming that pi follows Dirichlet(priorAlphas)
}
