% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liPosteriors.R
\name{postPii}
\alias{postPii}
\title{Compute M-H draw for pi_i in Li algorithm}
\usage{
postPii(sij, currentPii, priorA, priorB, S)
}
\arguments{
\item{sij}{Vector of cycle skip indicators for individual i}

\item{currentPii}{Current value of pi_i}

\item{priorA}{Hyperparameter alpha.}

\item{priorB}{Hyperparameter beta.}

\item{S}{Maximum number of skips allowed in algorithm}
}
\value{
Draw for pi_i, repeated for the number of observations from individual i
}
\description{
This performs a Metropolis-Hastings draw for pi_i, assuming s_ij follows a truncated geometric distribution with parameters
pi_i and S. The proposal distribution for pi_i is Beta(alpha, beta).
}
