% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postRho}
\alias{postRho}
\title{Sample a value from the full conditional posterior of rho}
\usage{
postRho(b)
}
\arguments{
\item{b}{Numeric vector, Random intercepts for individuals}
}
\value{
Numeric
}
\description{
In our model the data are drawn from LogN(mu_ij + log(c_ij), tau_i). mu_ij = Xi^T \%*\%Beta + bi, where the prior
for bi is given as N(0, rho). This function draws from the conditional posterior of rho, given
that the prior on rho is a uniform prior on the standard deviation.
}
