% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liPosteriors.R
\name{postSij}
\alias{postSij}
\title{Compute random draw from full conditional posterior for s_ij in Li algorithm.}
\usage{
postSij(yijs, pii, lambdai, S)
}
\arguments{
\item{yijs}{Vector of observed values for s_ij.}

\item{pii}{Probability parameter pi.}

\item{lambdai}{Value of lambda_i.}

\item{S}{Truncation level.}
}
\value{
A random draw from the posterior distribution of s_ij.
}
\description{
This function calculates a random draw from the full conditional posterior distribution
for s_ij in the Li algorithm, given the observed values yijs, the parameter pi,
the lambda_i value, and the truncation level S.
}
