% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postTauiOLD}
\alias{postTauiOLD}
\title{Sample a value from the full conditional posterior of tau_i (OLD VERSION)}
\usage{
postTauiOLD(yij, cij, mui, priorA = 1, priorB = 0.1)
}
\arguments{
\item{yij}{Numeric vector, cycle lengths for a single individual}

\item{cij}{Positive Integer vector, a sampled vector of length(yij) where the corresponding
values in cij indicate a sampled number of TRUE cycles in each cycle length given by yij}

\item{mui}{Numeric, log of sampled mean of individuals yijs}

\item{priorA}{Numeric > 0, prior shape of tau_i}

\item{priorB}{Numeric > 0, prior rate of tau_i}
}
\value{
Numeric
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_{ij}), tau_i). The prior for tau_i
is given as Gamma(priorA, priorB). This function draws from the conditional
posterior of tau_i. Note that we parameterize with RATE, not SCALE.
}
