% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.skipTrack.model}
\alias{print.skipTrack.model}
\title{Print skipTrack.model to console}
\usage{
\method{print}{skipTrack.model}(x, ...)
}
\arguments{
\item{x}{skipTrack.model object from the function skipTrack.fit}

\item{...}{Needed for S3 consistency}
}
\value{
Invisible NULL. Prints info about skipTrack.model object
}
\description{
Print skipTrack.model to console
}
