% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrackMCMC.R
\name{sampleStep}
\alias{sampleStep}
\title{Perform a single step of the MCMC sampling process.}
\usage{
sampleStep(
  ijDat,
  iDat,
  rho,
  pi,
  Xi,
  Zi,
  Beta,
  Gamma,
  priorAlphas,
  indFirst,
  rhoBeta,
  rhoGamma,
  phi
)
}
\arguments{
\item{ijDat}{A data.frame with individual-observation level parameters: Individual, ys, cs, muis, tauis.}

\item{iDat}{A data.frame with individual level parameters: Individual, mus, taus, thetas.}

\item{rho}{Updated value of the global parameter rho.}

\item{pi}{Updated value of the global parameter pi.}

\item{Xi}{A matrix of covariates for cycle length mean.}

\item{Zi}{A matrix of covariates for cycle length precision.}

\item{Beta}{Matrix of coefficients for cycle length mean.}

\item{Gamma}{Matrix of coefficients for cycle length precision.}

\item{priorAlphas}{Vector of prior alpha values for updating pi.}

\item{indFirst}{A logical vector indicating the first occurrence of each individual.}

\item{rhoBeta}{Updated value of the global parameter rhoBeta.}

\item{rhoGamma}{Updated value of the global parameter rhoGamma.}

\item{phi}{Value of the parameter phi.}
}
\value{
A list containing updated parameters after performing a single MCMC step.
The list includes:
\describe{
\item{ijDat}{A data.frame with updated parameters at the individual-observation level: Individual, ys, cs.}
\item{iDat}{A data.frame with updated parameters at the individual level: Individual, mus, taus.}
\item{rho}{Updated value of the global parameter rho.}
\item{pi}{Updated value of the global parameter pi.}
\item{Xi}{Matrix of covariates for cycle length mean.}
\item{Zi}{Matrix of covariates for cycle length precision.}
\item{Beta}{Matrix of coefficients for cycle length mean.}
\item{Gamma}{Matrix of coefficients for cycle length precision.}
\item{priorAlphas}{Vector of prior alpha values for updating pi.}
\item{indFirst}{A logical vector indicating the first occurrence of each individual.}
\item{rhoBeta}{Updated value of the global parameter rhoBeta.}
\item{rhoGamma}{Updated value of the global parameter rhoGamma.}
\item{phi}{Value of the parameter phi.}
}
}
\description{
This function performs a single step of the Markov Chain Monte Carlo (MCMC) algorithm to update parameters
in a hierarchical model used for identifying skips in menstrual cycle tracking.
}
