% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrack-package.R
\docType{package}
\name{skipTrack-package}
\alias{skipTrack}
\alias{skipTrack-package}
\title{skipTrack: Bayesian Hierarchical Models Adjusting for Non-Adherence in Mobile Menstrual Cycle Tracking}
\description{
skipTrack implements a Bayesian hierarchical model of the same name designed to perform inference on cycle length mean and regularity given the possibility of non-adherence in cycle length self-tracking. Currently accepts baseline continuous covariates for cycle mean length and regularity. Future updates will include include categorical covariates, time-varying covariates, and the inclusion of external information regarding tracking skips.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/skipTrack}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/skipTrack/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID})

}
\keyword{internal}
