% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{skipTrack.results}
\alias{skipTrack.results}
\title{Get a table of Inference results from skipTrack.fit}
\usage{
skipTrack.results(stFit, trueVals = NULL, burnIn = 750)
}
\arguments{
\item{stFit}{Object result of skipTrack.fit function.}

\item{trueVals}{Optional list containing true values for Betas, Gammas, and cijs.}

\item{burnIn}{Number of skipTrack iterations to discard as burn-in.}
}
\value{
A list containing the following elements:
\item{Betas}{95\% credible intervals and (if trueVals is supplied) true values for Betas and Coverage tag.}
\item{Gammas}{95\% credible intervals and (if trueVals is supplied) true values for Gammas and Coverage tag.}
\item{GammaGR}{Gelman-Rubin diagnostics for Gammas.}
\item{Cijs}{Wald confidence intervals and (if trueVals is supplied) true values for cijs and Coverage tags.}
\item{CijGR}{Gelman-Rubin diagnostics for cijs.}
}
\description{
This function calculates inference results on Betas, Gammas, and cijs based on the provided MCMC results.
It returns summaries such as quantiles and confidence intervals for Betas, Gammas, and cijs, as well as Gelman-Rubin diagnostics.
Note that true values and converage are included in the output if trueVals is supplied, but otherwise not.
}
\examples{
# Example usage with simulated data (which includes access to ground truth):
#
# skipTrack.results(stFitObject, trueVals = simulatedData, burnIn = 750)
#
# Example usage with real data:
#
# skipTrack.results(stFitObject, burnIn = 750)

}
