% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrackMCMC.R
\name{skipTrackMCMC}
\alias{skipTrackMCMC}
\title{Perform MCMC sampling for identifying skips in cycle tracking.}
\usage{
skipTrackMCMC(
  cycleDat,
  initialParams = list(pi = c(1/3, 1/3, 1/3), muis = rep(log(30),
    length(unique(cycleDat$Individual))), tauis = rep(5,
    length(unique(cycleDat$Individual))), mu = log(30), rho = 1, cs = rep(1,
    nrow(cycleDat))),
  reps = 1000
)
}
\arguments{
\item{cycleDat}{A data.frame with columns 'Individual' and 'TrackedCycles', representing the observed cycle lengths for each individual.}

\item{initialParams}{A list of initial parameter values for the MCMC algorithm.
Default values are provided for pi, muis, tauis, mu, rho, and cs.}

\item{reps}{The number of MCMC iterations (steps) to perform. Default is 1000.}
}
\value{
A list containing the MCMC draws for each parameter at each iteration. Each element
in the list is itself a list containing:
\describe{
\item{ijDat}{A data.frame with updated parameters at the individual-observation level: Individual, ys, cs.}
\item{iDat}{A data.frame with updated parameters at the individual level: Individual, mus, taus.}
\item{mu}{Updated value of the global parameter mu.}
\item{rho}{Updated value of the global parameter rho.}
\item{pi}{Updated value of the global parameter pi.}
\item{priorAlphas}{Vector of prior alpha values for updating pi.}
}
}
\description{
This function runs a Markov Chain Monte Carlo (MCMC) algorithm to update parameters in a hierarchical model (specified in...),
to identify skips in menstrual cycle tracking.
}
\examples{
# Example usage of skipTrackMCMC function
#COMMENTED TO SKIP DURING PACKAGE BUILD
#cycleDat <- simTrackData(1000, skipProb = c(.7, .2, .1))
#result <- skipTrackMCMC(cycleDat)

}
\seealso{
\code{\link{gibbsStep}}
}
\keyword{cycles}
\keyword{hierarchical}
\keyword{mcmc}
\keyword{model}
\keyword{skipped}
\keyword{tracking}
