% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str.skipTrack.model}
\alias{str.skipTrack.model}
\title{Report skipTrack.model structure to console}
\usage{
\method{str}{skipTrack.model}(object, ...)
}
\arguments{
\item{object}{skipTrack.model object from the function skipTrack.fit}

\item{...}{To match other str calls}
}
\value{
Invisible NULL. Prints structure of skipTrack.model object
}
\description{
Report skipTrack.model structure to console
}
\examples{
#Simulated data
simDat <- skipTrack.simulate(n = 100, skipProb = c(.7, .2, .1))

#Run model fit (should typically run with much more than 50 reps)
modFit <- skipTrack.fit(Y = simDat$Y, cluster = simDat$cluster, chains = 2, reps = 50)
str(modFit)


}
