% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{stResults}
\alias{stResults}
\title{Function to get inference results on Betas and Gammas and cijs}
\usage{
stResults(mcmcRes, dat = NULL, burnIn = 750)
}
\arguments{
\item{mcmcRes}{A list of MCMC results.}

\item{dat}{Optional data frame containing true values for Betas, Gammas, and cijs.}

\item{burnIn}{Number of iterations to discard as burn-in.}
}
\value{
A list containing the following elements:
\item{Betas}{Quantiles and (if dat is supplied) true values for Betas.}
\item{Gammas}{Quantiles and (if dat is supplied) true values for Gammas.}
\item{GammaGR}{Gelman-Rubin diagnostics for Gammas.}
\item{Cijs}{Confidence intervals and (if dat is supplied) true values for cijs.}
\item{CijGR}{Gelman-Rubin diagnostics for cijs.}
}
\description{
This function calculates inference results on Betas, Gammas, and cijs based on the provided MCMC results.
It returns summaries such as quantiles and confidence intervals for Betas, Gammas, and cijs, as well as Gelman-Rubin diagnostics.
Note that true values and converage are included in the output if dat is supplied, but otherwise not.
}
\examples{
# Example usage:
# stResults(mcmcRes, dat = my_data, burnIn = 750)

}
