% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{stSimCAT}
\alias{stSimCAT}
\title{Simulate user tracked menstrual cycle data for an individual, based on the skipTrack model WITH CATEGORICAL COVARIATES}
\usage{
stSimCAT(i, skipProb, maxCycles, trueBetas, trueGammas, overlap, avgCyclesPer)
}
\arguments{
\item{i}{Individual identifier. Character, numeric or integer.}

\item{skipProb}{Vector of probabilities for number of true cycles per tracked cycle. For
example, (.7, .2, .1) means that 70\% of observed cycles will contain one true cycle, 20\%
will contain 2 true cycles and 10\% will contain 3 true cycles.}

\item{maxCycles}{Maximum number of true cycles per tracked cycle. Ignored for this model.}

\item{trueBetas}{Optional. True values for the mean regression coefficients. Input as a list of vectors. Each list is one covariate, values of vectors give effects in that covariate.}

\item{trueGammas}{Optional. True values for the precision regression coefficients. Input as a list of vectors. Each list is one covariate, values of vectors give effects in that covariate.}

\item{overlap}{Optional. Number of (non-intercept) covariates shared between X and Z. Overlap is from top to bottom of list.}

\item{avgCyclesPer}{Average number of cycles contributed by each individual. Actual number is drawn from Poisson for each person. Default is 7.}
}
\value{
\describe{
\item{'Individual'}{Individual identifiers.}
\item{'TrackedCycles'}{Tracked cycles.}
\item{'NumTrue'}{Number of true values.}
\item{'LogMean'}{Individual's mean of log(Y).}
\item{'LogPrec'}{Individual's precision of log(Y)}
\item{'Beta0'}{Beta0 true value.}
\item{'Gamma0'}{Gamma0 true value.}
\item{'X0',...,'XN'}{Covariate matrix for Mean, where N is the length of trueBetas. Note that these will be integer values giving the category of the covariate.}
\item{'Z0',...,'ZM'}{Covariate matrix for precision, where M is the length of trueGammas. Note that these will be integer values giving the category of the covariate.}
}
}
\description{
This function generates synthetic data for user tracked menstrual cycles for a
single individual, where the covariates are categorical. For this model Beta_0 = log(30), Gamma_0 = 5.5, and phi = .01.
}
\seealso{
\code{\link{skipTrack.simulate}}
}
