% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrackData.R
\name{liSim}
\alias{liSim}
\title{Simulate user tracked menstrual cycle data for an individual using the li method.}
\usage{
liSim(i, skipProb, maxCycles, trueBetas, trueGammas = NULL)
}
\arguments{
\item{i}{Individual identifier. Character, numeric or integer.}

\item{skipProb}{Vector, ignored for this method.}

\item{maxCycles}{Integer, Maximum possible number of true cycles per tracked cycle.}

\item{trueBetas}{Optional. True values for generated mean regression coefficients.}

\item{trueGammas}{NULL, left for consistency. Will throw error if specified.}
}
\value{
A data.frame with columns 'Individual', 'TrackedCycles',
'NumTrue', 'Mean', 'SkipProb', 'Z0', 'Beta0', 'Gamma0', 'X0', ..., 'Xn' where n = length(trueBetas).
}
\description{
This function generates synthetic data for user tracked menstrual cycles for a
single individual using the li method.
}
\seealso{
\code{\link{simTrackData}}
}
