% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postBeta}
\alias{postBeta}
\title{Draw from Posterior Distribution for Beta Parameters}
\usage{
postBeta(rhoBeta = 0.01, rho, Xi, muI)
}
\arguments{
\item{rhoBeta}{A scalar representing the prior precision parameter for beta.}

\item{rho}{A scalar representing the precision parameter.}

\item{Xi}{A matrix of covariates, where each row represents an individual and each column represents a covariate.}

\item{muI}{A vector where each element is the mean for individual i.}
}
\value{
A vector representing a draw from the posterior distribution of beta parameters.
}
\description{
In our model mui follows a normal distribution with mean Xi^T \%*\% beta and precision rho. Additionally
we assume that beta follows a mvnormal prior with mean 0 and precision (rho_Beta) * I.This function draws
from the posterior distribution of beta under these assumptions.
}
\details{
This function assumes that \code{Xi} is a (num Individuals) x (dimension of beta) matrix of covariates.
}
