% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postCij}
\alias{postCij}
\title{Sample a value from the full conditional posterior of c_ij}
\usage{
postCij(yij1, pi, mui, taui)
}
\arguments{
\item{yij1}{Numeric, one cycle length for individual i}

\item{pi}{Numeric vector, must sum to 1. Sampled probabilities for c_ij}

\item{mui}{Numeric, log of sampled mean of individuals yijs}

\item{taui}{Numeric > 0, A sampled precision for the yijs}
}
\value{
Integer
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_{ij}), tau_i) The prior for
c_ij is a categorical prior with category probabilities pi1, ..., pik, and c_ij can
take values 1, ..., k where k is the length of pi. This function samples from the
full conditional posterior of c_ij
}
\examples{
postCij(yij1 = 60, pi = c(.7, .2, .1), mui = log(30), taui = 1)
}
