% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liPosteriors.R
\name{postLambdai}
\alias{postLambdai}
\title{Compute random draw from full conditional posterior for lambda_i in Li algorithm.}
\usage{
postLambdai(yij, sij, priorK, priorG)
}
\arguments{
\item{yij}{Vector of observed values for lambda_i.}

\item{sij}{Vector of indicators for lambda_i.}

\item{priorK}{Prior hyperparameter for K.}

\item{priorG}{Prior hyperparameter for G.}
}
\value{
A random draw from the posterior distribution of lambda_i.
}
\description{
This function calculates a random draw from the full conditional posterior distribution
for lambda_i in the Li algorithm, given the observed values y_ij, the indicators s_ij,
and the prior hyperparameters priorK and priorG.
}
\examples{
yij <- c(31, 62, 31)
sij <- c(0, 1, 0)
priorK <- 180
priorG <- 6
postLambdai(yij, sij, priorK, priorG)
}
