% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{postPhi}
\alias{postPhi}
\title{Metropolis-Hastings step to draw a new value for phi.}
\usage{
postPhi(taui, thetai, currentPhi, rhoPhi = 1000)
}
\arguments{
\item{taui}{Numeric vector, individuals precisions.}

\item{thetai}{Numeric vector. individuals precisions means (estimate)}

\item{currentPhi}{Previous draw of phi}

\item{rhoPhi}{Proposal rate for gamma distribution that draws proposal for phi, default is 1000.}
}
\value{
Numeric, new draw of phi
}
\description{
In our model the data are drawn from LogN(mu_i + log(c_ij), tau_i). The prior for tau_i
is given as Gamma(thetai\emph{phi, phi). This function uses a MH step to draw a new sample of phi.
Proposal distribution is Gamma(currentPhi}rhoPhi, rhoPhi).
Note that we parameterize with RATE, not SCALE.
}
