% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrack-package.R
\docType{package}
\name{skipTrack-package}
\alias{skipTrack}
\alias{skipTrack-package}
\title{skipTrack: A Bayesian Hierarchical Model that Controls for Non-Adherence in Mobile Menstrual Cycle Tracking}
\description{
Implements a Bayesian hierarchical model designed to identify skips in mobile menstrual cycle self-tracking on mobile apps. Future developments will allow for the inclusion of covariates affecting cycle mean and regularity, as well as extra information regarding tracking non-adherence. Main methods to be outlined in a forthcoming paper, with alternative models from Li et al. (2022) \doi{10.1093/jamia/ocab182}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/skipTrack}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/skipTrack/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID}) [copyright holder]

}
\keyword{internal}
