% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrack-package.R
\docType{package}
\name{skipTrack-package}
\alias{skipTrack}
\alias{skipTrack-package}
\title{skipTrack: Bayesian Hierarchical Models for Identifying Non-Adherence in Mobile Menstrual Cycle Tracking}
\description{
skipTrack implements an in-development Bayesian hierarchical model designed to estimate the posterior probability that a given logged menstrual cycle in a mobile app contains 1 or more true cycles. Currently only uses menstrual cycle length data, will be expanded later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/skipTrack}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/skipTrack/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID})

}
\keyword{internal}
