% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{skipTrack.diagnostics}
\alias{skipTrack.diagnostics}
\title{Takes model results from skipTrack.fit and uses genMCMCDiag to get mcmc diagnostics.}
\usage{
skipTrack.diagnostics(stFit, param, method = NULL, ...)
}
\arguments{
\item{stFit}{A list of MCMC results from the skipTrack.fit function.}

\item{param}{A character string specifying the parameter for which diagnostics are to be calculated.
Must be one of: 'mu', 'rho', 'muis', 'tauis', or 'cijs'.}

\item{method}{An optional parameter specifying the method for calculating diagnostics. Default is NULL.}

\item{...}{
  Arguments passed on to \code{\link[genMCMCDiag:genDiagnostic]{genMCMCDiag::genDiagnostic}}
  \describe{
    \item{\code{diagnostics}}{A character vector or list of diagnostic functions to be evaluated. Options include 'traceplot', 'ess', 'gelmanRubin', or custom functions. See details.}
    \item{\code{distance}}{Function for evaluating distance between MCMC draws if required by 'method'. Note that the lanfear and ts methods ALWAYS require a distance function.}
    \item{\code{verbose}}{If TRUE, informative messages are displayed.}
  }}
}
\value{
A mcmcDiag object of MCMC diagnostics for the specified parameter
}
\description{
Takes model results from skipTrack.fit and uses genMCMCDiag to get mcmc diagnostics.
}
\details{
If the parameter is 'rho' (the univariate parameters),
the function extracts the specified parameter from the MCMC results and calculates
diagnostics using the genDiagnostic function with the
standard method. If the parameter is 'cijs', 'muis', or 'tauis', the
function extracts the corresponding values and calculates diagnostics using the genDiagnostic
function with the specified or default method ('lanfear') and hammingDist as the distance function.
}
\examples{
#Example usage:
#skipTrack.diagnostics(stFitObject, 'mu')

}
\seealso{
\code{\link{genDiagnostic}}, \code{\link{skipTrack.fit}}
}
