% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrackMulti.R
\name{skipTrack.fit}
\alias{skipTrack.fit}
\title{Runs the skipTrack Model, using multiple MCMC chains}
\usage{
skipTrack.fit(
  Y,
  cluster,
  X = matrix(1, nrow = length(unique(cluster))),
  Z = matrix(1, nrow = length(unique(cluster))),
  numSkips = 10,
  reps = 1000,
  chains,
  useParallel = TRUE,
  li = FALSE,
  liHyperparams = c(kappa = 180, gamma = 6, alpha = 2, beta = 20),
  ...
)
}
\arguments{
\item{Y}{A vector of observed cycle lengths.}

\item{cluster}{A vector indicating the individual cluster/group membership for each observation.}

\item{X}{A matrix (numIndividuals x length(Beta)) of covariates for cycle length mean. Default is a vector of 1's.}

\item{Z}{A matrix (numIndividuals x length(Gamma)) of covariates for cycle length precision. Default is a vector of 1's.}

\item{numSkips}{The maximum number of skips to allow. Default is 10.}

\item{reps}{The number of MCMC iterations (steps) to perform. Default is 1000.}

\item{chains}{Number of chains to run in parallel or sequentially.}

\item{useParallel}{Logical, indicating whether to use parallel processing. Default is TRUE.}

\item{li}{Logical, if TRUE runs the estimation given the model provided in Li et al. (2022).
This model does not estimate covariates, so these will be ignored.}

\item{liHyperparams}{Named numeric vector, the hyperparameters named in the Li algorithm. Defaults
are provided, will only be used if li == TRUE}

\item{...}{
  Arguments passed on to \code{\link[=skipTrackMCMC]{skipTrackMCMC}}
  \describe{
    \item{\code{fixedSkips}}{If TRUE cycle skip information (cijs) is not updated in sample steps and the inputs are instead assumed to be true.}
    \item{\code{initialParams}}{A list of initial parameter values for the MCMC algorithm.
Default values are provided for pi, muis, tauis, mu, rho, cs, alphas, Beta, Gamma, phi, rhoBeta, rhoGamma, and rhoPhi.}
  }}
}
\value{
A list containing the results of skipTrackMCMC for each chain.
}
\description{
This function runs skipTrackMCMC on multiple chains,
either in parallel or sequentially.
If li == TRUE can perform estimation using the model given in Li et al. (2022).
}
\seealso{
\code{\link{skipTrackMCMC}}
}
