% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{skipTrack.visualize}
\alias{skipTrack.visualize}
\title{Visualize Results from skipTrack.fit}
\usage{
skipTrack.visualize(mcmcRes)
}
\arguments{
\item{mcmcRes}{A list containing MCMC results obtained from skipTrack.fit.}
}
\value{
A list of three ggplot2 objects:
\itemize{
\item{cijOverPlt}{Scatter plot of estimated Cij values against reported cycle length.}
\item{cijOverTaus}{Scatter plot of estimated Cij values against estimated individual precisions, colored by cycle length.}
\item{muByChain}{Density plot of the sampled density of overall mean by chain.}
}
}
\description{
This function takes results from skipTrack.fit and produces helpful visualizations of the results.
}
\examples{
# Example usage:
# skipTrack.visualize(stFitObject)

}
\seealso{
\code{\link{skipTrack.fit}} for generating MCMC results to be visualized.
}
