% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipTrackMCMC.R
\name{skipTrackMCMC}
\alias{skipTrackMCMC}
\title{Perform MCMC sampling to identify skips in cycle tracking.}
\usage{
skipTrackMCMC(
  Y,
  cluster,
  X = matrix(1, nrow = length(unique(cluster))),
  Z = matrix(1, nrow = length(unique(cluster))),
  numSkips = 10,
  reps = 1000,
  fixedSkips = FALSE,
  initialParams = list(pi = rep(1/(numSkips + 1), numSkips + 1), muis = rep(log(30),
    length(unique(cluster))), tauis = rep(5, length(unique(cluster))), rho = 1, cs =
    sample(1:3, length(Y), replace = TRUE), alphas = rep(1, numSkips + 1), Beta =
    matrix(rep(0, ncol(as.matrix(X))), 1), Gamma = matrix(rep(0, ncol(as.matrix(Z))), 1),
    rhoBeta = 1, rhoGamma = 1000, phi = 0.01)
)
}
\arguments{
\item{Y}{A vector of observed cycle lengths.}

\item{cluster}{A vector indicating the individual cluster/group membership for each observation.}

\item{X}{A matrix of covariates for cycle length mean. Default is a matrix of 1's.}

\item{Z}{A matrix of covariates for cycle length precision. Default is a matrix of 1's.}

\item{numSkips}{The maximum number of skips to allow. Default is 10.}

\item{reps}{The number of MCMC iterations (steps) to perform. Default is 1000.}

\item{fixedSkips}{If TRUE cycle skip information (cijs) is not updated in sample steps and the inputs are instead assumed to be true.}

\item{initialParams}{A list of initial parameter values for the MCMC algorithm.
Default values are provided for pi, muis, tauis, mu, rho, cs, alphas, Beta, Gamma, rhoBeta, rhoGamma, and phi.}
}
\value{
A list containing the MCMC draws for each parameter at each iteration. Each element
in the list is itself a list containing:
\describe{
\item{ijDat}{A data.frame with updated parameters at the individual-observation level: Individual, ys, cs.}
\item{iDat}{A data.frame with updated parameters at the individual level: Individual, mus, taus.}
\item{rho}{Updated value of the global parameter rho.}
\item{pi}{Updated value of the global parameter pi.}
\item{Xi}{Matrix of covariates for cycle length mean.}
\item{Zi}{Matrix of covariates for cycle length precision.}
\item{Beta}{Matrix of coefficients for cycle length mean.}
\item{Gamma}{Matrix of coefficients for cycle length precision.}
\item{priorAlphas}{Vector of prior alpha values for updating pi.}
\item{indFirst}{A logical vector indicating the first occurrence of each individual.}
\item{rhoBeta}{Updated value of the global parameter rhoBeta.}
\item{rhoGamma}{Updated value of the global parameter rhoGamma.}
\item{phi}{Value of the parameter phi.}
\item{fixedSkips}{Logical. Indicates if skips were fixed.}
}
}
\description{
This function runs a Markov Chain Monte Carlo (MCMC) algorithm to update parameters in a hierarchical model
to identify skips in menstrual cycle tracking.
}
\examples{
# Example usage of skipTrackMCMC function
cycleDat <- simTrackData(10, skipProb = c(.7, .2, .1)) #Simulated Data
result <- skipTrackMCMC(cycleDat)

#MORE REALISTIC VERSION
#COMMENTED TO SKIP DURING PACKAGE BUILD
#cycleDat <- simTrackData(1000, skipProb = c(.7, .2, .1)) #Simulated Data
#result <- skipTrackMCMC(cycleDat)

}
\seealso{
\code{\link{gibbsStep}}
}
\keyword{cycles}
\keyword{hierarchical}
\keyword{mcmc}
\keyword{model}
\keyword{skipped}
\keyword{tracking}
