% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str.skipTrack.model}
\alias{str.skipTrack.model}
\title{Report skipTrack.model structure to console}
\usage{
\method{str}{skipTrack.model}(x)
}
\arguments{
\item{x}{skipTrack.model object from the function skipTrack.fit}
}
\value{
Invisible NULL. Prints structure of skipTrack.model object
}
\description{
Report skipTrack.model structure to console
}
